ARCHIVE/requirements.txt
colorama==0.4.4
PyYAML==6.0
termcolor==1.1.0